+++ cypress/integration/circulation-log/search.spec.js
@@ -6,13 +6,13 @@ import permissions from '../../support/dictionary/permissions';
 import UsersSearchPane from '../../support/fragments/users/usersSearchPane';
 import UsersCard from '../../support/fragments/users/usersCard';
 import CheckinActions from '../../support/fragments/check-in-actions/checkInActions';
-import CheckoutActions from '../../support/fragments/checkout/checkout';
 import InventoryHoldings from '../../support/fragments/inventory/holdings/inventoryHoldings';
 import devTeams from '../../support/dictionary/devTeams';
 import Users from '../../support/fragments/users/users';
 import UserEdit from '../../support/fragments/users/userEdit';
 import ServicePoints from '../../support/fragments/settings/tenant/servicePoints/servicePoints';
 import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
+import Checkout from '../../support/fragments/checkout/checkout';
 
 const ITEM_BARCODE = `123${getRandomPostfix()}`;
 let userId;
@@ -73,7 +73,7 @@ describe('circulation-log', () => {
             });
           })
           .then(() => {
-            CheckoutActions.checkoutItemViaApi({
+            Checkout.checkoutItemViaApi({
               itemBarcode: ITEM_BARCODE,
               userBarcode: Cypress.env('users')[0].barcode,
               servicePointId,
