+++ cypress/integration/requests/requests-sort.spec.js
@@ -11,8 +11,8 @@ describe('ui-requests: Sort requests', () => {
   const instances = [];
 
   beforeEach(() => {
-    cy.login(Cypress.env('diku_login'), Cypress.env('diku_password'));
-    cy.getToken('diku_admin', 'admin');
+    cy.loginAsAdmin({ path: TopMenu.requestsPath, waiter: Requests.waitContentLoading });
+    cy.getAdminToken();
 
     Object.values(Requests.requestTypes).forEach((requestType) => {
       const itemStatus = requestType === 'Page' ? 'Available' : 'Checked out';
