--- cypress/support/fragments/check-out/confirmMultiplePiecesItemCheckOut.js
@@ -1,7 +1,0 @@
-import { Button, Modal } from '../../../../interactors';
-
-export default {
-  confirmMultiplePiecesItemModal:() => {
-    cy.do(Modal('Confirm multipiece check out').find(Button('Check out')).click());
-  },
-};
